% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bn.fit.hurdle}
\alias{bn.fit.hurdle}
\title{bn.fit.hurdle}
\usage{
bn.fit.hurdle(x, data, cdrAdjustment = FALSE)
}
\arguments{
\item{x}{bn object}

\item{data}{data to be fitted}

\item{cdrAdjustment}{adjust for cellular detection rate based on `data`}
}
\description{
make a pseudo bn.fit object using continuous part of Hurdle model.
}
\details{
You should not use `rbn` or relevant functions for logic sampling 
to this object as  the logistic regression part of the model will 
not be taken into account.
}
\examples{
data(gaussian.test)
test.bn <- hc(gaussian.test)
bn.fit.hurdle(test.bn, gaussian.test)
}
