% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bn_fit_to_igraph}
\alias{bn_fit_to_igraph}
\title{bn_fit_to_igraph}
\usage{
bn_fit_to_igraph(graph)
}
\arguments{
\item{graph}{bn.fit object}
}
\description{
This function coverts the bn.fit object to igraph object.
The nodes with no parents or nodes will not be included in the resulting igraph.
The coefficient value is stored in `coef` attribute of resulting graph.
}
