% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc.auprc}
\alias{calc.auprc}
\title{calc.auprc
calculate the AUPRC based on reference BN and inferred network.
The function assumes the directed network.}
\usage{
calc.auprc(ref, str, target = "strength", mode = "directed")
}
\arguments{
\item{ref}{reference bn object or adjacency matrix}

\item{str}{inferred strength or weight (three-column with from, to, and target column)}

\item{target}{target column name}

\item{mode}{directed or undirected}
}
\description{
calc.auprc
calculate the AUPRC based on reference BN and inferred network.
The function assumes the directed network.
}
