% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccdrboot.R
\name{ccdr.boot}
\alias{ccdr.boot}
\title{ccdr.boot}
\usage{
ccdr.boot(
  data,
  R = 200,
  m = nrow(data),
  lambdas.length = 20,
  alpha = 2,
  gamma = 2
)
}
\arguments{
\item{alpha}{passed to ccdr.run}

\item{gamma}{passed to ccdr.run}

\item{lambda.length}{passed to generate.lambdas}
}
\description{
Bootstrap-based arc strength calculation based on CCDr algorithm.
The function uses ccdrAlgorithm::ccdr.run for the sampled data, and returns the 
bootstrap strength for the arcs per lambda.
}
