% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_tbl_graph}
\alias{convert_to_tbl_graph}
\title{convert_to_tbl_graph
convert bn object to tbl_graph object}
\usage{
convert_to_tbl_graph(bn)
}
\arguments{
\item{bn}{bn object}
}
\value{
tbl_graph
}
\description{
convert_to_tbl_graph
convert bn object to tbl_graph object
}
