% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_metrics.R
\name{evaluate_metrics}
\alias{evaluate_metrics}
\title{evaluate_metrics}
\usage{
evaluate_metrics(
  fitted,
  N,
  pens = c("glmnet_CV"),
  orig = TRUE,
  alphas = c(0.001, 0.005, 0.01, 0.05),
  ccdr = TRUE,
  return_net = FALSE
)
}
\description{
s0: edge number in original graph
p: node number in original graph
N: sample number
}
