% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{gen}
\alias{gen}
\title{gen
random.graph based on gene node name}
\usage{
gen(n, data, method = "melancon", params = list())
}
\arguments{
\item{method}{ic-dag, melancon or ordered. default to melancon.}

\item{params}{pass to random.graph}
}
\value{
bn object and bn.fit object
}
\description{
gen
random.graph based on gene node name
}
