% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ges.pcalg.R
\name{ges.pcalg}
\alias{ges.pcalg}
\title{ges.pcalg
Perform Greedy Equivalent Search implemented in pcalg}
\usage{
ges.pcalg(data, args = NULL)
}
\arguments{
\item{data}{data}

\item{args}{arguments to pcalg::ges}
}
\description{
ges.pcalg
Perform Greedy Equivalent Search implemented in pcalg
}
