% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenes.R
\name{getGOGenes}
\alias{getGOGenes}
\title{getGOGenes}
\usage{
getGOGenes(GO, orgDb = org.Mm.eg.db, type = "SYMBOL")
}
\arguments{
\item{GO}{GO ID}

\item{orgDb}{AnnotationDbi that has "GOALL" slot}

\item{type}{return type, default to SYMBOL}
}
\description{
get the vector of genes from GO term
}
