% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadppi.R
\name{getKEGGedges}
\alias{getKEGGedges}
\title{returnKEGGedges}
\usage{
getKEGGedges(pathID, args = list(), largestComponents = TRUE, bn = TRUE)
}
\arguments{
\item{pathID}{pathway ID}

\item{args}{arguments of ggkegg::pathway}

\item{bn}{if TRUE, try to convert to bn object, but if failed, return igraph}
}
\description{
return directed igraph object
}
\details{
this function tries to extract DAG from KEGG PATHWAY,
and not necessarily succeed.

The function uses ggkegg to obtain and parse KEGG PATHWAY information
using KEGG RESTful API.  There should be type `gene` in the node data of pathway
this returns the directed relationship described in the pathway
}
