% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenes.R
\name{getKEGGPathwayGenes}
\alias{getKEGGPathwayGenes}
\title{getKEGGPathwayGenes}
\usage{
getKEGGPathwayGenes(pid, org = "mmu", orgDb = org.Mm.eg.db, type = "SYMBOL")
}
\arguments{
\item{pid}{pathway id in KEGG}

\item{org}{organism name}

\item{orgDb}{organism database}

\item{type}{returned ID type of genes}
}
\description{
get genes related to KEGG PATHWAY
}
