% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeletons.R
\name{glmnetBICpath}
\alias{glmnetBICpath}
\title{glmnetBICpath
return glmnet BIC path for lambdas}
\usage{
glmnetBICpath(data, nn)
}
\arguments{
\item{data}{data for fitting}

\item{nn}{node name for Y}
}
\description{
glmnetBICpath
return glmnet BIC path for lambdas
}
\examples{
library(ggplot2)
data(gaussian.test)
glmnetBICpath(gaussian.test, "A")$plot
}
