% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalStruc.R
\name{globalStruc}
\alias{globalStruc}
\title{globalStruc}
\usage{
globalStruc(
  spe,
  candidate_genes,
  label,
  exclude_label = NA,
  algorithm = "mmhc",
  reg = FALSE,
  algorithm.args = list(),
  return_bn = FALSE,
  return_data = FALSE,
  debug = FALSE,
  cluster_label = NULL,
  penalty = "glmnet"
)
}
\arguments{
\item{algorithm}{inference algorithm of BN, ignored if reg=TRUE}

\item{cluster_labels}{(experimental) aggregate to speed up computation}
}
\description{
globalStruc
}
