% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalStrucValues.R
\name{globalStrucValues}
\alias{globalStrucValues}
\title{globalStrucValues}
\usage{
globalStrucValues(
  spe,
  global_tbl_graph = NULL,
  label,
  exclude_label = NA,
  summarize_func = mean,
  variation_func = sd,
  bn = NULL
)
}
\description{
globalStrucValues
}
