% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeletons.R
\name{hurdle.bic}
\alias{hurdle.bic}
\title{Custom score function for bnlearn, using hurdle model in glmmTMB.
BIC is scaled by -2.
Custom score function for bnlearn, using hurdle model.}
\usage{
hurdle.bic(node, parents, data, args)
}
\arguments{
\item{node}{candidate node}

\item{parents}{parents of the nodes}

\item{data}{data used for scoring}

\item{args}{other arguments}
}
\description{
Custom score function for bnlearn, using hurdle model in glmmTMB.
BIC is scaled by -2.
Custom score function for bnlearn, using hurdle model.
}
