% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadppi.R
\name{intersectPpi}
\alias{intersectPpi}
\title{intersectPpi}
\usage{
intersectPpi(edge_names, org = "mm", return_net = FALSE)
}
\arguments{
\item{edge_names}{edge names in `GeneA->GeneB` style}

\item{org}{organism name (mm, or hsa)}
}
\description{
Examine intersection of PPI.
Named vector of edge number, intersection number and proportion will be returned.
}
