% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interVal.R
\name{interVal}
\alias{interVal}
\title{interVal}
\usage{
interVal(
  data,
  algos = c("hc", "mmhc"),
  algorithm.args = list(list(), list()),
  mode = "shd",
  r = 10,
  ss = 100,
  verbose = FALSE,
  returnA0 = FALSE
)
}
\arguments{
\item{ss}{sample size}
}
\description{
interVal
}
