% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcalgFuns.R
\name{lingam.pcalg}
\alias{lingam.pcalg}
\title{lingam.pcalg
Perform LiNGAM implemented in pcalg}
\usage{
lingam.pcalg(data, args = NULL)
}
\arguments{
\item{data}{data}

\item{args}{arguments to pcalg::ges}
}
\description{
lingam.pcalg
Perform LiNGAM implemented in pcalg
}
