% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerCoefs.R
\name{markerCoefs}
\alias{markerCoefs}
\title{markerCoefs}
\usage{
markerCoefs(
  coef_mat,
  classif_label = "group",
  cell_label = NULL,
  cell_column = "label",
  sample_column = "Sample",
  tentative_fix = TRUE,
  return_mat = FALSE,
  verbose = FALSE,
  returnChar = TRUE,
  xgboost = FALSE,
  xgboostArgs = list()
)
}
\arguments{
\item{coef_mat}{returned value of `strucValues`}

\item{classif_label}{which label to be used for classification}

\item{cell_label}{cell label for fitting}

\item{cell_column}{which column represents cell types}

\item{sample_column}{sample column in colData()}

\item{tentative_fix}{if Boruta, perform tentative fix 
if unconfirmed features are present}

\item{return_mat}{return matrix used for classification}

\item{verbose}{control logging}

\item{returnChar}{return character vector of confirmed features}

\item{xgboost}{perform xgboost instead of Boruta}

\item{xgboostArgs}{parameters passed to xgboost function}
}
\description{
identify marker regulatory relationships based on 
fitted coefficient matrix.
}
\details{
Users can choose XGBoost or Boruta algorithm.
}
