% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtainSubNetMatrix.R
\name{obtainSubNetMatrix}
\alias{obtainSubNetMatrix}
\title{obtainSubNetMatrix}
\usage{
obtainSubNetMatrix(df, spe, func = "sum")
}
\arguments{
\item{df}{the results of `globalStrucValues` (specifying the bn object)}
}
\description{
obtainSubNetMatrix
}
