% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcalgFuns.R
\name{pcalg.boot.future}
\alias{pcalg.boot.future}
\title{pcalg.boot.future}
\usage{
pcalg.boot.future(
  data,
  R = 200,
  m = nrow(data),
  removeAllZero = FALSE,
  verbose = FALSE,
  return.all = FALSE,
  fun = "ges",
  args = NULL
)
}
\arguments{
\item{data}{data (row: sample, column: gene)}

\item{R}{replicate number}

\item{m}{sampling number}

\item{removeAllZero}{remove all zero genes per replicate}

\item{verbose}{control verbosity}

\item{return.all}{return all the network in bootstrapping}

\item{fun}{pcalg functions, lingam or ges}

\item{args}{args to be passed to the main function of pcalg}
}
\value{
bn.strength
}
\description{
Bootstrap-based arc strength calculation based on LiNGAM and GES.
The future.lapply will be used in this function.
}
