% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perLabelStruc.R
\name{perLabelStruc}
\alias{perLabelStruc}
\title{perLabelStruc}
\usage{
perLabelStruc(
  spe,
  label,
  candidate_genes,
  algorithm = "mmhc",
  reg = TRUE,
  penalty = "glmnet",
  use_assay = "logcounts",
  all = FALSE,
  label_name = NULL,
  verbose = FALSE,
  algorithm.args = list(),
  barcode_column = "Barcode",
  change_symbol = TRUE,
  symbol_column = "Symbol",
  cluster_label = NULL
)
}
\arguments{
\item{spe}{SingleCellExperiment or SpatialExperiment}

\item{all}{infer all the network of the corresponding label}

\item{label_name}{if all=FALSE, must be specified}
}
\description{
Infer the BN per label (like using only the Epithelial cells or one sample)
The input is SingleCellExperiment or SpatialExperiment object and candidate genes to be included in the inference,
and label column in colData.
}
