% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perLabelStrucValues.R
\name{perLabelStrucValues}
\alias{perLabelStrucValues}
\title{perLabelStrucValues}
\usage{
perLabelStrucValues(strucs, remove_zero = TRUE)
}
\arguments{
\item{strucs}{named list of bn.strength}
}
\description{
perLabelStrucValues
}
