% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plasso.R
\name{plasso.fit.single}
\alias{plasso.fit.single}
\title{plasso.fit.single}
\usage{
plasso.fit.single(
  X,
  y,
  lambda = 1,
  eps = 1e-06,
  P = NULL,
  maxIter = 100,
  gamma = 0.5,
  lr = 1e-06,
  tol = 1e-05,
  mu = 0.01
)
}
\arguments{
\item{X}{matrix}

\item{y}{response vector}
}
\description{
reimplementation of PrecisionLasso.
Currently, only the continuous response is supported.
Automatic determination of gamma will be supported.
RcppArmadillo version is faster (plasso_fit).
}
