% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.prc}
\alias{plot.prc}
\title{plot.prc
Plot the PRC based on reference BN and inferred network.
The function assumes the directed network.}
\usage{
\method{plot}{prc}(ref.bn, strs, target = "strength", onlyData = FALSE)
}
\arguments{
\item{ref.bn}{reference bn object}

\item{target}{target column name (must be same in all the str)}

\item{onlyData}{return only the data}

\item{str}{list of inferred strength or weight (three-column with from, to, and target column)}
}
\description{
plot.prc
Plot the PRC based on reference BN and inferred network.
The function assumes the directed network.
}
