% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAVN.R
\name{plotAVN}
\alias{plotAVN}
\title{plotAVN}
\usage{
plotAVN(
  str,
  edge = geom_edge_diagonal,
  ew = strength,
  lyt = "fr",
  size_degree = TRUE,
  size_range = c(1, 5)
)
}
\arguments{
\item{ew}{edge width, strength or direction}
}
\description{
plotAVN
}
