% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNet.R
\name{plotNet}
\alias{plotNet}
\title{plotNet
plot inferred network (with coefficients if available)}
\usage{
plotNet(
  net,
  data = NULL,
  layout = "kk",
  geom = geom_edge_link,
  largest_components = TRUE
)
}
\arguments{
\item{net}{network in bn class}
}
\description{
plotNet
plot inferred network (with coefficients if available)
}
