% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAVNPerCell.R
\name{plotNetPerCell}
\alias{plotNetPerCell}
\title{plotNetPerCell}
\usage{
plotNetPerCell(
  net = NULL,
  avn = NULL,
  str = NULL,
  values,
  edge = geom_edge_diagonal,
  lyt = "fr",
  size_degree = TRUE
)
}
\arguments{
\item{values}{the data.frame output of globalStrucValues(), storing per-cell (per-cluster) parameters.
If needed, values should be subset beforehand for the interesting clusters.}
}
\description{
plotNetPerCell
}
