% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSTNet.R
\name{plotSTNet}
\alias{plotSTNet}
\title{plotSTNet}
\usage{
plotSTNet(
  spe,
  ig,
  label = "label",
  edge_width = NULL,
  edge_color = NULL,
  edge_range = c(0.5, 1.5),
  sample_id = NULL,
  image_id = NULL,
  return_tbl_graph = FALSE,
  point_label = "ground_truth",
  cell_cluster_label = NULL,
  directed = TRUE,
  use_ggfx = TRUE,
  point_size = 2,
  rep_colour = "black",
  edge_base_colour = "grey50",
  edge_base_width = 0.5
)
}
\arguments{
\item{spe}{SpatialExperiment}

\item{ig}{igraph or data.frame object containing `from` and `to` column}

\item{label}{label in `colData(spe)` that matches the node labels in ig}

\item{edge_width}{edge width in igraph or data.frame}

\item{edge_color}{edge color in igraph or data.frame}

\item{edge_range}{edge width range}

\item{sample_id}{sample_id in `spe`, used to subset image}

\item{image_id}{image_id in `spe`, used to subset image}

\item{return_tbl_graph}{return tbl graph only}

\item{point_label}{background point label}

\item{cell_cluster_label}{label of each node, should be named vector}

\item{directed}{network is directed or not}

\item{use_ggfx}{use ggfx or not}

\item{point_size}{representative point size}

\item{rep_colour}{representative point colour}

\item{edge_base_colour}{fixed edge colour}

\item{edge_base_width}{fixed edge width}
}
\value{
ggplot
}
\description{
Overlay igraph network on histology image
}
