% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSTWithNum.R
\name{plotSTWithNum}
\alias{plotSTWithNum}
\title{plotSTWithNum}
\usage{
plotSTWithNum(
  spe,
  num,
  label = "label",
  sample_id = NULL,
  image_id = NULL,
  point_size = 1.5
)
}
\arguments{
\item{spe}{SpatialExperiment}

\item{num}{named vector of num}

\item{label}{label in `colData(spe)` that matches the node labels in ig}

\item{sample_id}{sample_id in `spe`, used to subset image}

\item{image_id}{image_id in `spe`, used to subset image}

\item{point_size}{point size}
}
\value{
ggplot
}
\description{
Overlay histology images by using named vector of values
}
