% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSTWithSummarized.R
\name{plotSTWithSummarized}
\alias{plotSTWithSummarized}
\title{plotSTWithSummarized}
\usage{
plotSTWithSummarized(
  spe,
  gsv,
  from_node,
  to_node,
  label = "label",
  sample_id = NULL,
  image_id = NULL
)
}
\arguments{
\item{spe}{SpatialExperiment}

\item{gsv}{tbl_graph object}

\item{from_node}{from node}

\item{to_node}{to node}

\item{label}{label in `colData(spe)` that matches the node labels in ig}

\item{sample_id}{sample_id in `spe`, used to subset image}

\item{image_id}{image_id in `spe`, used to subset image}
}
\value{
ggplot
}
\description{
Overlay igraph network on histology image
}
