% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSubNet.R
\name{plotSubNet}
\alias{plotSubNet}
\title{plotSubNet}
\usage{
plotSubNet(
  df,
  candidate_node_id,
  sort = FALSE,
  with_hist = FALSE,
  spe = NULL,
  label = "label",
  node_size = 3,
  show = NULL,
  layout = "kk"
)
}
\arguments{
\item{df}{the results of `globalStrucValues`}

\item{candidate_node_id}{candidate node id to subset}
}
\description{
plotSubNet
}
