% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVisiumWithNum.R
\name{plotVisiumWithNum}
\alias{plotVisiumWithNum}
\title{plotSTNet}
\usage{
plotVisiumWithNum(
  spe,
  gsv,
  from_node,
  to_node,
  label = "label",
  sample_id = NULL,
  image_id = NULL
)
}
\arguments{
\item{spe}{SpatialExperiment}

\item{label}{label in `colData(spe)` that matches the node labels in ig}

\item{sample_id}{sample_id in `spe`, used to subset image}

\item{image_id}{image_id in `spe`, used to subset image}

\item{ig}{igraph or data.frame object containing `from` and `to` column}

\item{edge_width}{edge width in igraph or data.frame}

\item{edge_color}{edge color in igraph or data.frame}

\item{edge_range}{edge width range}

\item{return_tbl_graph}{return tbl graph only}

\item{point_label}{background point label}

\item{cell_cluster_label}{label of each node, should be named vector}

\item{directed}{network is directed or not}

\item{use_ggfx}{use ggfx or not}

\item{point_size}{representative point size}

\item{rep_colour}{representative point colour}

\item{edge_base_colour}{fixed edge colour}

\item{edge_base_width}{fixed edge width}
}
\value{
ggplot
}
\description{
Overlay igraph network on histology image
}
