% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prc.plot}
\alias{prc.plot}
\title{prc.plot
Plot the PRC based on reference BN and inferred network.
The function assumes the directed network.}
\usage{
prc.plot(ref.bn, strs, target = "strength", onlyData = FALSE)
}
\arguments{
\item{ref.bn}{reference bn object}

\item{strs}{list of inferred strength or weight (three-column with from, to, and target column)}

\item{target}{target column name (must be same in all the str)}

\item{onlyData}{return only the data}
}
\description{
prc.plot
Plot the PRC based on reference BN and inferred network.
The function assumes the directed network.
}
