% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadppi.R
\name{returnKEGGedges}
\alias{returnKEGGedges}
\title{returnKEGGedges}
\usage{
returnKEGGedges(pathID, ggkeggArgs)
}
\arguments{
\item{pathID}{pathway ID}

\item{ggkeggArgs}{arguments of ggkegg::pathway}
}
\description{
return directed igraph object
}
\details{
The function uses ggkegg to obtain and parse KEGG PATHWAY information
using KEGG RESTful API.  There should be type `gene` in the node data of pathway
this returns the directed relationship described in the pathway
}
