% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scstruc.R
\name{scstruc}
\alias{scstruc}
\title{scstruc}
\usage{
scstruc(sce, path)
}
\arguments{
\item{sce}{class object storing single cell data}

\item{path}{pathway ID}
}
\description{
Main function wrapping the structure learning and fitting.
}
