% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sel.genes.kegg.ora}
\alias{sel.genes.kegg.ora}
\title{sel.genes.kegg.ora
select genes based on KEGG over-representation analysis}
\usage{
sel.genes.kegg.ora(
  genes,
  orgDb = org.Hs.eg.db,
  keyType = "ENSEMBL",
  sel = NULL,
  organism = "hsa"
)
}
\description{
sel.genes.kegg.ora
select genes based on KEGG over-representation analysis
}
