% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shdmat}
\alias{shdmat}
\title{shdmat}
\usage{
shdmat(netlist)
}
\arguments{
\item{netlist}{list of bn}
}
\description{
shdmat
}
\examples{
data(gaussian.test)
test <- head(gaussian.test, 10)
shdmat(list(hc(test), mmhc(test)))
}
