% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sid.bn.R
\name{sid.bn}
\alias{sid.bn}
\title{sid.bn}
\usage{
sid.bn(true.bn, est.bn, twoway = FALSE)
}
\arguments{
\item{true.bn}{true bn object}

\item{est.bn}{estimated bn object}

\item{twoway}{return the sid(true.bn, est.bn) + sid(est.bn, true.bn) divided by 2}
}
\description{
given the bn object, calculate SID between them.
}
