% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeletons.R
\name{skeleton.reg.boot}
\alias{skeleton.reg.boot}
\title{skeleton.reg.boot}
\usage{
skeleton.reg.boot(
  data,
  algorithm = "glmnet_CV",
  R = 100,
  m = nrow(data),
  algorithm.args = list()
)
}
\arguments{
\item{data}{input data}

\item{algorithm}{two-stage algorithm}

\item{R}{replicate number}

\item{m}{sampling number}

\item{algorithm.args}{passed to algorithm function}
}
\description{
bootstrapping the two-stage approaches
}
