% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacells.R
\name{superCellMat}
\alias{superCellMat}
\title{superCellMat}
\usage{
superCellMat(
  GE,
  pca = FALSE,
  genes = NULL,
  gamma = 10,
  k.knn = 5,
  mode = "average"
)
}
\arguments{
\item{GE}{gene (row) to cell (column) matrix, like log-normalized}

\item{pca}{use PCA or not}

\item{genes}{subset of the genes to use in PCA (`getTopHVGs`, etc)}

\item{gamma}{param to SCimplify}

\item{k.knn}{param to SCimplify}

\item{mode}{average or sum}
}
\description{
Return the metacell abundance matrix
}
