% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add.dropout}
\alias{add.dropout}
\title{add.dropout
Like Splat and SERGIO, add dropout to matrix}
\usage{
add.dropout(mat, shape = 6.5, q = 0.65)
}
\value{
binary matrix indicating which cell to be zero-ed out
}
\description{
add.dropout
Like Splat and SERGIO, add dropout to matrix
}
