% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bn_fit_to_igraph}
\alias{bn_fit_to_igraph}
\title{bn_fit_to_igraph}
\usage{
bn_fit_to_igraph(graph, preserve = FALSE)
}
\arguments{
\item{graph}{bn.fit object}

\item{preserve}{preserve the node with no parents and children}
}
\description{
This function coverts the bn.fit object to igraph object.
The nodes with no parents and children will not be included in the resulting igraph.
The coefficient value is stored in `coef` attribute of resulting graph.
}
\examples{
data(gaussian.test)
test <- head(gaussian.test)
struc <- hc(test)
bn_fit_to_igraph(bn.fit(struc, test))
}
