% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc.auprc}
\alias{calc.auprc}
\title{calc.auprc
calculate the AUPRC based on reference BN and inferred network.
The function assumes the directed network.}
\usage{
calc.auprc(ref.bn, str, target = "strength")
}
\arguments{
\item{ref.bn}{reference bn object}

\item{str}{inferred strength or weight (three-column with from, to, and target column)}

\item{target}{target column name}
}
\description{
calc.auprc
calculate the AUPRC based on reference BN and inferred network.
The function assumes the directed network.
}
