% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_metrics.R
\name{evaluate_metrics}
\alias{evaluate_metrics}
\title{evaluate_metrics}
\usage{
evaluate_metrics(
  fitted,
  N,
  pens = c("glmnet_CV"),
  mmhc = TRUE,
  alphas = c(0.001, 0.005, 0.01, 0.05),
  ppi = FALSE,
  database = "string",
  org = "mm",
  ccdr = TRUE,
  return_net = FALSE,
  lambdas.length = 20,
  sid = FALSE
)
}
\arguments{
\item{fitted}{reference network that is parameter fitted}

\item{N}{number of samples to be drawn from `fitted` by `rbn` function.}

\item{mmhc}{perform comparison with MMHC implemented in bnlearn with
alphas specified in `alphas` argument.}

\item{ppi}{if specified, include metrics of comparison with PPI.
The network node name should be symbol.}

\item{ccdr}{perform comparison with CCDr algorithm with lambda length specified in 
`lambdas.length` argument.}

\item{sid}{compute SID, needs package SID in CRAN.}
}
\description{
s0: edge number in original graph
edges: edge number in inferred graph
p: node number in original graph
N: sample number
SHD: structural hamming distance
SID: structural intervention distance (if specified)
time: time needed in structure learning (sec)
}
\details{
Time will be calculated by differences in Sys.time()
}
