% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractOmnipath.R
\name{extractDAGfromOmnipath}
\alias{extractDAGfromOmnipath}
\title{extractDAGfromOmnipath
This function takes the pathway name in SignaLink and
1. Subset the pathway-related genes in dorothea interaction and remove duplicates
2. Identify and remove minimal feedback arc set (make DAG)
3. Takes the largest components in the graph}
\usage{
extractDAGfromOmnipath(
  genes = NULL,
  pathway_name = NULL,
  target_source = "AND",
  mfas = TRUE
)
}
\arguments{
\item{mfas}{force DAG}
}
\description{
extractDAGfromOmnipath
This function takes the pathway name in SignaLink and
1. Subset the pathway-related genes in dorothea interaction and remove duplicates
2. Identify and remove minimal feedback arc set (make DAG)
3. Takes the largest components in the graph
}
\examples{
extractDAGfromOmnipath("TGF")
}
