% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{gen}
\alias{gen}
\title{gen}
\usage{
gen(n, data, method = "melancon", params = list())
}
\arguments{
\item{n}{sample number}

\item{data}{data to be fitted}

\item{method}{arguments in bnlearn::random.graph().
ic-dag, melancon or ordered. default to melancon.}

\item{params}{pass to random.graph}
}
\value{
bn object and bn.fit object
}
\description{
generate random.graph based on gene node name and returns the 
fitted network based on the user-provided data
}
