% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKEGGEdges.R
\name{getKEGGEdges}
\alias{getKEGGEdges}
\title{returnKEGGedges}
\usage{
getKEGGEdges(
  pathID,
  args = list(),
  largestComponents = TRUE,
  bn = TRUE,
  removeCycle = FALSE
)
}
\arguments{
\item{pathID}{pathway ID}

\item{args}{arguments of ggkegg::pathway}

\item{largestComponents}{obtain largest components in the graph}

\item{bn}{if TRUE, try to convert to bn object, but if failed, return igraph}

\item{removeCycle}{remove cycle using feedback_arc_set()}
}
\description{
return directed igraph object
}
\details{
The function uses ggkegg to obtain and parse KEGG PATHWAY information
using KEGG RESTful API.  There should be type `gene` in the node data of pathway
this returns the directed relationship described in the pathway.
The function tries to extract DAG from KEGG PATHWAY, and not necessarily succeed.
The first part of graphics name in the graph is taken as the representative gene name.
}
