% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggraph.compare.R
\name{ggraph.compare}
\alias{ggraph.compare}
\title{ggraph.compare
compare two bn object based on parallel edges}
\usage{
ggraph.compare(nets, lyt = "kk")
}
\arguments{
\item{nets}{(named) list of bn object}

\item{layout}{layout in ggraph}
}
\value{
ggplot2 object
}
\description{
ggraph.compare
compare two bn object based on parallel edges
}
