% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalStruc.R
\name{globalStruc}
\alias{globalStruc}
\title{globalStruc}
\usage{
globalStruc(
  spe,
  candidate_genes,
  label = NULL,
  algorithm = "mmhc",
  algorithm.args = list(),
  return_bn = TRUE,
  return_data = TRUE,
  cluster_label = NULL,
  verbose = FALSE,
  use_assay = "logcounts",
  barcode_column = "row",
  change_symbol = TRUE,
  symbol_column = "Symbol",
  input = NULL,
  nonzero = 1,
  rmNeg = FALSE
)
}
\arguments{
\item{algorithm}{inference algorithm of BN
If you want to use CCDr algorithm, specify `ccdr` or `ccdr.boot`.
If you use CCDr algorithm, the {bn} object is returned.}

\item{use_assay}{use assay, default to logcounts}

\item{input}{default to NULL}

\item{cluster_labels}{(experimental) aggregate to speed up computation}
}
\description{
globalStruc
}
