% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalStruc.R
\name{globalStruc}
\alias{globalStruc}
\title{globalStruc}
\usage{
globalStruc(
  spe,
  candidate_genes,
  label,
  algorithm = "mmhc",
  reg = FALSE,
  algorithm.args = list(),
  return_bn = FALSE,
  return_data = FALSE,
  cluster_label = NULL,
  penalty = "glmnet",
  verbose = FALSE,
  use_assay = "logcounts",
  barcode_column = "Barcode",
  change_symbol = TRUE,
  symbol_column = "Symbol"
)
}
\arguments{
\item{algorithm}{inference algorithm of BN, ignored if reg=TRUE}

\item{penalty}{if {reg}=TRUE, the penalty used can be specified, default to {glmnet} (lasso).
If you want to use CCDr algorithm, specify `ccdr` or `ccdr.boot`.
If you use CCDr algorithm, the {bn} object is returned.}

\item{use_assay}{use assay}

\item{cluster_labels}{(experimental) aggregate to speed up computation}
}
\description{
globalStruc
}
