% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdle.boot.R
\name{hurdle.boot}
\alias{hurdle.boot}
\title{hurdle.boot}
\usage{
hurdle.boot(
  data,
  R = 200,
  m = nrow(data),
  score = NULL,
  removeAllZero = FALSE,
  debug = FALSE
)
}
\arguments{
\item{data}{data (row: sample, column: gene)}

\item{R}{replicate number}

\item{m}{sampling number}

\item{score}{scoring function after identification of skeleton.}

\item{removeAllZero}{remove all zero genes in replicates}

\item{debug}{passed to maximize function}
}
\value{
bn.strength
}
\description{
Bootstrap-based arc strength calculation based on Hurdle model.
}
