% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interVal.R
\name{interVal}
\alias{interVal}
\title{interVal}
\usage{
interVal(
  data,
  algos = c("hc", "mmhc"),
  algorithm.args = list(list(), list()),
  mode = "shd",
  r = 10,
  ss = 100,
  verbose = FALSE,
  returnA0 = FALSE
)
}
\arguments{
\item{data}{data (row names as genes and columns as samples)}

\item{algos}{algorithms to be evaluated.}

\item{algorithm.args}{list of lists to be passed to algos}

\item{mode}{shd or sid}

\item{r}{number of iteration}

\item{ss}{sample size}

\item{verbose}{control verbosity}

\item{returnA0}{returns the A0 (intersection of inferred networks)}
}
\value{
summarized statistics
}
\description{
Performs `Intersection-Validation` approach to the data.
}
\examples{
data(gaussian.test)
interVal(head(gaussian.test))
}
