% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interVal.R
\name{interVal}
\alias{interVal}
\title{interVal}
\usage{
interVal(
  data,
  algos = c("hc", "mmhc"),
  algorithm.args = NULL,
  r = 10,
  ss = 100,
  verbose = FALSE,
  returnA0 = TRUE,
  symmetrize = FALSE,
  SID.cran = FALSE
)
}
\arguments{
\item{data}{data (row names as genes and columns as samples)}

\item{algos}{algorithms to be evaluated.}

\item{algorithm.args}{list of lists to be passed to algos}

\item{r}{number of iteration}

\item{ss}{sample size}

\item{verbose}{control verbosity}

\item{returnA0}{returns the A0 (intersection of inferred networks), default to TRUE}

\item{symmetrize}{use symmetrized version of SID}

\item{SID.cran}{use SID in CRAN package SID}
}
\value{
summarized statistics, bn object of A0, and statistics for all the subsamples
}
\description{
Performs `Intersection-Validation` approach to the data.
}
\examples{
data(gaussian.test)
interVal(head(gaussian.test, n=50), ss=10)
}
