% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\title{metrics}
\usage{
metrics(reference, inferred)
}
\arguments{
\item{reference}{reference bn object}

\item{inferred}{named list of bn object}
}
\value{
data.frame storing evaluation results
}
\description{
Calculate and return metrics
}
