% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\title{metrics}
\usage{
metrics(reference, inferred, sid_sym = FALSE)
}
\arguments{
\item{reference}{reference bn object}

\item{inferred}{named list of bn object}

\item{sid_sym}{symmetric version of sid will be computed}
}
\value{
data.frame storing evaluation results
}
\description{
Calculate and return metrics
}
\examples{
library(bnlearn)
data(gaussian.test)
ref <- hc(head(gaussian.test))
metrics(ref, list("ref1"=ref,"ref2"=ref))
}
