% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtainSubNetMatrix.R
\name{obtainSubNetMatrix}
\alias{obtainSubNetMatrix}
\title{obtainSubNetMatrix}
\usage{
obtainSubNetMatrix(df, spe, func = "sum", useAssay = "logcounts")
}
\arguments{
\item{df}{the results of `strucValues` (specifying the bn object)}

\item{spe}{SingleCellExperiment object}

\item{func}{how to summarize subnetwork abundance}

\item{useAssay}{which assay to use}
}
\description{
obtain subnetwork abundance matrix
}
