% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAVN.R
\name{plotAVN}
\alias{plotAVN}
\title{plotAVN}
\usage{
plotAVN(
  str,
  edge = geom_edge_link,
  ew = strength,
  layout = "kk",
  degreeMode = "all",
  sizeDegree = TRUE,
  sizeRange = c(1, 5)
)
}
\arguments{
\item{str}{boot.strength object}

\item{edge}{geom to be used in edge}

\item{ew}{edge width, strength or direction}

\item{layout}{layout in ggraph}

\item{degreeMode}{degree mode in igraph::degree}

\item{sizeDegree}{size the node based on degree}

\item{sizeRange}{node size range}
}
\description{
plot the averaged network based on bootstrapping
}
