% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAVN.R
\name{plotAVN}
\alias{plotAVN}
\title{plotAVN}
\usage{
plotAVN(avn, str, edge = geom_edge_diagonal, ew = strength, lyt = "fr")
}
\arguments{
\item{ew}{edge width, strength or direction}
}
\description{
plotAVN
}
