% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAVN.R
\name{plotAVN}
\alias{plotAVN}
\title{plotAVN}
\usage{
plotAVN(
  str,
  edge = geom_edge_link,
  ew = "strength",
  ec = "direction",
  layout = "kk",
  degreeMode = "all",
  sizeDegree = TRUE,
  sizeRange = c(3, 6),
  data = NULL,
  highlightColor = "tomato",
  highlightExpand = 5,
  highlightEdges = NULL,
  largestComponents = TRUE,
  threshold = NULL
)
}
\arguments{
\item{str}{boot.strength object}

\item{edge}{geom to be used in edge}

\item{ew}{edge width, strength, direction, or coef
if coef, data to be fitted must be provided}

\item{ec}{edge color, strength, direction, or coef
if coef, data to be fitted must be provided}

\item{layout}{layout in ggraph}

\item{degreeMode}{degree mode in igraph::degree}

\item{sizeDegree}{size the node based on degree}

\item{sizeRange}{node size range}

\item{data}{data to fit the paramter on averaged network}

\item{highlightColor}{highlight color for ggfx}

\item{highlightExpand}{expand parameter for ggfx}

\item{highlightEdges}{highlight edges, two-column matrix is expected
If highlight is specified, ggfx::with_outer_glow 
will be used to highlight the edge}

\item{largestComponents}{take largest components automatically}

\item{threshold}{threshold for averaged network}
}
\description{
plot the averaged network based on bootstrapping
}
