% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNetPerCell.R
\name{plotNetPerCell}
\alias{plotNetPerCell}
\title{plotNetPerCell}
\usage{
plotNetPerCell(
  values,
  net = NULL,
  avn = NULL,
  str = NULL,
  lyt = "fr",
  size_degree = TRUE
)
}
\arguments{
\item{values}{the data.frame output of strucValues(), storing per-cell (per-cluster) parameters.
If needed, values should be subset beforehand for the interesting clusters.}

\item{net}{igraph or bn object}

\item{avn}{averaged network}

\item{str}{strength object}

\item{lyt}{layout}

\item{size_degree}{whether to size the nodes using the centrality_degree()}
}
\description{
The function is for plotting the coefficient inside the graph per cell.
}
