% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotST.R
\name{plotST}
\alias{plotST}
\title{plotST}
\usage{
plotST(
  spe,
  label = "label",
  sample_id = NULL,
  image_id = NULL,
  point_size = 1.5
)
}
\arguments{
\item{spe}{SpatialExperiment}

\item{label}{label in `colData(spe)`}

\item{sample_id}{sample_id in `spe`, used to subset image}

\item{image_id}{image_id in `spe`, used to subset image}

\item{point_size}{point size}
}
\value{
ggplot
}
\description{
Overlay histology images
}
\examples{
# library(SpatialExperiment)
# example(read10xVisium, echo = FALSE)
# plotST(spe, "in_tissue")
}
