% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSubNet.R
\name{plotSubNet}
\alias{plotSubNet}
\title{plotSubNet}
\usage{
plotSubNet(
  df,
  candidate_node_id,
  sort = FALSE,
  with_hist = FALSE,
  spe = NULL,
  label = "label",
  node_size = 3,
  show = NULL,
  layout = "kk",
  cell_label = NULL,
  cell_column = NULL,
  highlight_edge = NULL,
  coef_cutoff = 0.01
)
}
\arguments{
\item{df}{the results of `strucValues`}

\item{candidate_node_id}{candidate node id to subset}

\item{sort}{sort the label}

\item{with_hist}{histology image plotting}

\item{spe}{SpatialExperiment class object}

\item{label}{name in spe}

\item{node_size}{node size}

\item{show}{which label to show}

\item{layout}{layout in ggraph}

\item{cell_label}{cell label name in cell_column}

\item{cell_column}{cell column name}

\item{highlight_edge}{edge highlighting}

\item{coef_cutoff}{coefficient cutoff to be plotted}
}
\value{
ggplot
}
\description{
plot subnetwork centered to specified gene
}
