% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSubNet.R
\name{plotSubNet}
\alias{plotSubNet}
\title{plotSubNet}
\usage{
plotSubNet(
  df,
  candidate_node_id,
  sort = FALSE,
  with_hist = FALSE,
  spe = NULL,
  label = "label",
  node_size = 3,
  show = NULL,
  layout = "kk",
  cell_label = NULL,
  cell_column = NULL,
  highlight_edge = NULL,
  coef_cutoff = 0.01
)
}
\arguments{
\item{df}{the results of `globalStrucValues`}

\item{candidate_node_id}{candidate node id to subset}

\item{cell_label}{if you need to plot per cell basis, specify cell label,
along with cell column}

\item{cell_column}{if you need to plot per cell basis, specify cell column,
along with cell label}

\item{coef_cutoff}{edge with value below this thrshold will not be drawn.}
}
\description{
plotSubNet
}
