% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scstruc.R
\name{scstruc}
\alias{scstruc}
\title{scstruc}
\usage{
scstruc(
  spe,
  candidateGenes,
  label = NULL,
  labelName = NULL,
  algorithm = "mmhc",
  algorithm.args = list(),
  returnBn = TRUE,
  returnData = TRUE,
  clusterLabel = NULL,
  verbose = FALSE,
  useAssay = "logcounts",
  barcodeColumn = "row",
  changeSymbol = TRUE,
  symbolColumn = "Symbol",
  input = NULL,
  nonzero = 1,
  rmNeg = FALSE
)
}
\arguments{
\item{spe}{SingleCellExperiment, SpatialExpetiment object}

\item{label}{label column}

\item{algorithm}{inference algorithm of BN
If you want to use CCDr algorithm, specify `ccdr` or `ccdr.boot`.
If you use CCDr algorithm, the {bn} object is returned.}

\item{returnData}{By default, return the data used for the inference.}

\item{useAssay}{use assay, default to logcounts}

\item{barcodeColumn}{barcode for cell identification}

\item{input}{default to NULL}

\item{clusterLabels}{(experimental) aggregate the count across this cluster to speed up computation}

\item{returBn}{By default, returns the bn class object}
}
\description{
The main function for scstruc package.
}
