% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scstruc.R
\name{scstruc}
\alias{scstruc}
\title{scstruc}
\usage{
scstruc(
  sce,
  candidateGenes,
  label = NULL,
  labelName = NULL,
  algorithm = "mmhc",
  algorithm.args = list(),
  returnBn = TRUE,
  returnData = TRUE,
  clusterLabel = NULL,
  verbose = FALSE,
  useAssay = "logcounts",
  changeSymbol = TRUE,
  symbolColumn = "Symbol",
  input = NULL,
  zeroFilt = 1,
  rmNeg = FALSE,
  returnOnlyData = FALSE,
  boot = FALSE,
  R = 200,
  m = NULL
)
}
\arguments{
\item{sce}{SingleCellExperiment, SpatialExpetiment object}

\item{candidateGenes}{candidate genes to be used.}

\item{label}{label column}

\item{labelName}{label name used to subset the cells}

\item{algorithm}{inference algorithm of BN
If you want to use CCDr algorithm, specify `ccdr` or `ccdr.boot`.
If you use CCDr algorithm, the \code{bn} object is returned.}

\item{algorithm.args}{list of algorithm arguments}

\item{returnBn}{By default, returns the bn class object}

\item{returnData}{By default, return the data used for the inference.}

\item{clusterLabel}{(experimental) aggregate the count across this cluster to speed up computation}

\item{verbose}{used for logging.}

\item{useAssay}{use assay, default to logcounts}

\item{changeSymbol}{automatically change the node name of input matrix to gene symbol}

\item{symbolColumn}{if changeSymbol is TRUE, use `symbolColumn` in rowData of the input object.}

\item{input}{default to NULL, if specified use this input matrix for inference, ignoring the `sce`.}

\item{zeroFilt}{parameter to filter the genes with many zeros.
Genes with the sum of zero expression cells <= the number of all cells * `zeroFilt`
will be used as input. By default perform no filtering (1).
This filter will be performed *after* the genes and cells subset by the other parameters.}

\item{rmNeg}{remove genes with all negative abundances}

\item{returnOnlyData}{return only the data, not performing inference.}

\item{boot}{bootstrapping specification}

\item{R}{when boot is TRUE, use as replicate number}

\item{m}{sampling number in bootstrapping}
}
\description{
The main function for scstruc package.
Perform input subset, structure learning.
}
\details{
\code{algorithm}: inference algorithm (two-stage or one-stage)
if \code{Hurdle} algorithm is specified, it returns the list of \code{hurdle}, \code{undir}, \code{bn}.
\code{hurdle}: returned object of HurdleNormal::fitHurdle.
\code{undir}: Undirected network with the best (lowest) BIC.
\code{bn}: bn object
if \code{ccdr}, the list of network for multiple lambdas will be returned
}
\examples{
library(scran)
library(scstruc)
sce <- mockSCE()
sce <- logNormCounts(sce)
included_genes <- sample(row.names(sce), 10)
gs <- scstruc(sce, included_genes, changeSymbol=FALSE)
gs.glmnet <- scstruc(sce, included_genes, changeSymbol=FALSE, algorithm="glmnet_BIC")
}
