% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sel.genes.kegg.ora}
\alias{sel.genes.kegg.ora}
\title{sel.genes.kegg.ora
select genes based on KEGG over-representation analysis}
\usage{
sel.genes.kegg.ora(
  genes,
  orgDb = org.Hs.eg.db,
  keyType = "ENSEMBL",
  sel = NULL,
  organism = "hsa"
)
}
\arguments{
\item{genes}{query genes}

\item{orgDb}{organism database object}

\item{keyType}{key type of genes}

\item{sel}{select the number of pathways, sorted by p-value}

\item{organism}{organism ID in KEGG, default to hsa}
}
\description{
sel.genes.kegg.ora
select genes based on KEGG over-representation analysis
}
