% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sidmat}
\alias{sidmat}
\title{sidmat}
\usage{
sidmat(netlist)
}
\arguments{
\item{netlist}{list of bn}
}
\description{
sidmat
}
\examples{
data(gaussian.test)
test <- head(gaussian.test, 10)
sidmat(list(hc(test), mmhc(test)))
}
