% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton_reg.R
\name{skeleton.reg}
\alias{skeleton.reg}
\title{skeleton.reg}
\usage{
skeleton.reg(
  data,
  penalty = "glmnet_CV",
  whitelist = NULL,
  blacklist = NULL,
  nFolds = 5,
  verbose = FALSE
)
}
\description{
Two-stage approach for regularization-based structure learning.
Currently, this function is experimental as 
it uses non-exported functions in `bnlearn`.
}
