% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacells.R
\name{superCellMat}
\alias{superCellMat}
\title{superCellMat}
\usage{
superCellMat(
  sce,
  genes = NULL,
  prop = 0.2,
  pca = TRUE,
  gamma = 10,
  k.knn = 5,
  rank = 10,
  mode = "average",
  ID = "ID",
  verbose = TRUE
)
}
\arguments{
\item{genes}{subset of the genes to use in PCA (`getTopHVGs`, etc)}

\item{pca}{use PCA or not}

\item{gamma}{param to SCimplify}

\item{k.knn}{param to SCimplify}

\item{mode}{average or sum}

\item{GE}{gene (row) to cell (column) matrix, like log-normalized}
}
\description{
Return the metacell abundance matrix
Should provide SingleCellExperiment with logcounts data filled.
}
